+++ lib/swid/frontend/dart/narrowDartTypeToSwidType.dart
