+++ lib/swid/frontend/dart/swidFunctionTypeFromFunctionType.dart
@@ -9,7 +9,7 @@ import 'package:hydro_sdk/swid/ir/swidType.dart';
 import 'package:hydro_sdk/swid/ir/swidTypeFormal.dart';
 import 'package:hydro_sdk/swid/ir/util/mapAnalyzerNullabilitySuffix.dart';
 import 'package:hydro_sdk/swid/ir/util/mapClassLibrarySourcePath.dart';
-import 'package:hydro_sdk/swid/ir/util/narrowDartTypeToSwidType.dart';
+import 'package:hydro_sdk/swid/frontend/dart/narrowDartTypeToSwidType.dart';
 
 import 'package:analyzer/dart/element/type.dart'
     show FunctionType, InterfaceType;
