+++ lib/swid/frontend/dart/swidInterfaceFromInterface.dart
@@ -6,7 +6,7 @@ import 'package:meta/meta.dart';
 import 'package:hydro_sdk/swid/ir/swidReferenceDeclarationKind.dart';
 import 'package:hydro_sdk/swid/ir/util/mapAnalyzerNullabilitySuffix.dart';
 import 'package:hydro_sdk/swid/ir/util/mapClassLibrarySourcePath.dart';
-import 'package:hydro_sdk/swid/ir/util/narrowDartTypeToSwidType.dart';
+import 'package:hydro_sdk/swid/frontend/dart/narrowDartTypeToSwidType.dart';
 
 import 'package:analyzer/dart/element/type.dart' show InterfaceType;
 import 'package:hydro_sdk/swid/ir/swidInterface.dart';
